requirements.txt
google-genai
streamlit
pillow
python-dotenv
pytest
altair<5
